+++ orbit/pkg/packaging/linux_shared.go
@@ -40,7 +40,7 @@ func buildNFPM(opt Options, pkger nfpm.Packager) (string, error) {
 
 	updateOpt.RootDirectory = orbitRoot
 	updateOpt.Targets = update.LinuxTargets
-	updateOpt.ServerCertificate = opt.UpdateTLSServerCertificate
+	updateOpt.ServerCertificatePath = opt.UpdateTLSServerCertificate
 
 	if opt.UpdateTLSClientCertificate != "" {
 		updateClientCrt, err := tls.LoadX509KeyPair(opt.UpdateTLSClientCertificate, opt.UpdateTLSClientKey)
