+++ pkg/certificate/certificate.go
@@ -5,10 +5,12 @@ import (
 	"context"
 	"crypto/tls"
 	"crypto/x509"
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"net"
 	"net/url"
+	"os"
 
 	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 )
@@ -117,4 +119,101 @@ func getHostPort(u *url.URL) string {
 		return net.JoinHostPort(host, u.Scheme)
 	}
 	return net.JoinHostPort(host, port)
+}
+
+// Certificate holds a loaded TLS certificate and its raw parts.
+type Certificate struct {
+	Crt    tls.Certificate
+	RawCrt []byte
+	RawKey []byte
+}
+
+// LoadCertificateFromFiles loads a TLS certificate from PEM cert and key file paths.
+//
+// Returns (nil, nil) if both files do not exist.
+func LoadCertificateFromFiles(crtPath, keyPath string) (*Certificate, error) {
+	checkFileExists := func(filePath string) (bool, error) {
+		switch s, err := os.Stat(filePath); {
+		case err == nil:
+			return !s.IsDir(), nil
+		case errors.Is(err, os.ErrNotExist):
+			return false, nil
+		default:
+			return false, err
+		}
+	}
+
+	crtExists, err := checkFileExists(crtPath)
+	if err != nil {
+		return nil, err
+	}
+	keyExists, err := checkFileExists(keyPath)
+	if err != nil {
+		return nil, err
+	}
+
+	if crtExists != keyExists {
+		return nil, fmt.Errorf(
+			"both crt and key files must exist: %s: %t, %s: %t",
+			crtPath, crtExists, keyPath, keyExists,
+		)
+	}
+	if !crtExists {
+		return nil, nil
+	}
+
+	crtBytes, err := os.ReadFile(crtPath)
+	if err != nil {
+		return nil, err
+	}
+	keyBytes, err := os.ReadFile(keyPath)
+	if err != nil {
+		return nil, err
+	}
+	crt, err := parseFullCertificate(crtBytes, keyBytes)
+	if err != nil {
+		return nil, err
+	}
+
+	return &Certificate{
+		Crt:    crt,
+		RawCrt: crtBytes,
+		RawKey: keyBytes,
+	}, nil
+}
+
+// LoadCertificate loads a certificate from the given PEM cert and key strings.
+//
+// Returns (nil, nil) if both values are empty.
+func LoadCertificate(crt, key string) (*tls.Certificate, error) {
+	if (crt != "") != (key != "") {
+		return nil, fmt.Errorf(
+			"both crt and key must be set: crt=%t, key=%t", crt != "", key != "",
+		)
+	}
+	if crt == "" {
+		return nil, nil
+	}
+
+	cert, err := parseFullCertificate([]byte(crt), []byte(key))
+	if err != nil {
+		return nil, err
+	}
+
+	return &cert, nil
+}
+
+func parseFullCertificate(crt, key []byte) (tls.Certificate, error) {
+	cert, err := tls.X509KeyPair(crt, key)
+	if err != nil {
+		return tls.Certificate{}, err
+	}
+	// tls.X509KeyPair does not store the parsed certificate leaf.
+	// To reduce per-handshake processing, we parse it here.
+	parsedLeaf, err := x509.ParseCertificate(cert.Certificate[0])
+	if err != nil {
+		return tls.Certificate{}, fmt.Errorf("parse leaf certificate: %w", err)
+	}
+	cert.Leaf = parsedLeaf
+	return cert, nil
 }
